Fixed package structure (#42)
* Fixed package structure

Added docs to PrefixAwareTextTokeniser.java warning that it is experimental
Restructured usecase package to separate subpackages (lookup and search)

* Removed separate builder classes for LocationService.java and SearchLocationService.java
