+++ location4j/src/main/java/com/tomaytotomato/location4j/usecase/lookup/FindState.java
@@ -1,11 +1,11 @@
-package com.tomaytotomato.location4j.usecase;
+package com.tomaytotomato.location4j.usecase.lookup;
 
 import com.tomaytotomato.location4j.model.State;
 import java.util.List;
 import java.util.Optional;
 
 /**
- * Interface for finding states based on various criteria.
+ * Interface for finding States based on various known criteria such as ID and Name.
  */
 public interface FindState {
