+++ location4j/src/main/java/com/tomaytotomato/location4j/usecase/search/SearchLocation.java
@@ -1,4 +1,4 @@
-package com.tomaytotomato.location4j.usecase;
+package com.tomaytotomato.location4j.usecase.search;
 
 import com.tomaytotomato.location4j.model.Location;
 import java.util.List;
