+++ location4j/src/test/java/com/tomaytotomato/location4j/usecase/FindCountryTest.java
@@ -4,9 +4,9 @@ package com.tomaytotomato.location4j.usecase;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.tomaytotomato.location4j.LocationService;
+import com.tomaytotomato.location4j.usecase.lookup.LocationService;
 import com.tomaytotomato.location4j.model.Country;
-import com.tomaytotomato.location4j.usecase.FindCountry;
+import com.tomaytotomato.location4j.usecase.lookup.FindCountry;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
