+++ location4j/src/test/java/com/tomaytotomato/location4j/usecase/FindStateTest.java
@@ -1,9 +1,10 @@
 package com.tomaytotomato.location4j.usecase;
 
 
-import com.tomaytotomato.location4j.LocationService;
+import com.tomaytotomato.location4j.usecase.lookup.LocationService;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import com.tomaytotomato.location4j.usecase.lookup.FindState;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
